pyproject.toml
[tool.poetry]
name = "innohassle-events"
version = "0.1.0"
description = ""
authors = ["Ruslan Bel'kov <ruslan.belckov@yandex.ru>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "~3.11"
sqlalchemy = {extras = ["postgresql_asyncpg"], version = "^2.0.17"}
pydantic = {extras = ["dotenv"], version = "^1.10.9"}
fastapi = {extras = ["all"], version = "^0.98.0"}
authlib = "^1.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
sqlalchemy2-stubs = "^0.0.2a34"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
